Remove redundant comments and import stataments
