+++ lessons/helpers.py
@@ -1,8 +1,6 @@
-from .models import Request, Booking, User
+from .models import Request, Booking
 from django.conf import settings
 from django.shortcuts import redirect
-from django.contrib import messages
-from django.urls import reverse
 
 def get_requests(user):
     # To do: Change to user when implemented
@@ -23,16 +21,6 @@ def get_all_bookings():
     # To do: Change to user when implemented
     bookings = Booking.objects.filter()
     return bookings
-
-def admin_required(view_function):
-    def modified_view_function(request, *args, **kwargs):
-        if request.user.is_student:
-            messages.add_message(request, messages.ERROR, "You cannot access this page!")
-            return redirect(settings.REDIRECT_URL_WHEN_LOGGED_IN)
-        else:
-            return view_function(request, *args, **kwargs)
-    return modified_view_function
-
 
 def login_prohibited(view_function):
     def modified_view_function(request):
